refactor: Domian 클래스로 변경(MenuProducts, Products)
