+++ src/main/java/kitchenpos/menus/domain/MenuChangePriceService.java
@@ -2,9 +2,8 @@ package kitchenpos.menus.domain;
 
 import kitchenpos.menus.domain.exception.InvalidMenuProductsPriceException;
 import kitchenpos.menus.domain.exception.InvalidMenuProductsSizeException;
-import kitchenpos.menus.domain.vo.MenuProducts;
 import kitchenpos.products.application.ProductService;
-import kitchenpos.products.domain.vo.Products;
+import kitchenpos.products.domain.Products;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
